Remove readings file
