+++ measurementcache.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"log"
 	"sort"
+	"sync"
 	"time"
 
 	. "github.com/gonium/gosdm630/internal/meters"
@@ -69,9 +70,9 @@ func (mc *MeasurementCache) Purge(deviceId byte) error {
 	if item, ok := mc.items[deviceId]; ok {
 		item.readings.Purge(deviceId)
 		return nil
-	} else {
-		return fmt.Errorf("No device with id %d available.", deviceId)
 	}
+
+	return fmt.Errorf("No device with id %d available.", deviceId)
 }
 
 func (mc *MeasurementCache) GetSortedIDs() []byte {
@@ -83,7 +84,7 @@ func (mc *MeasurementCache) GetSortedIDs() []byte {
 	return keys
 }
 
-func (mc *MeasurementCache) GetLast(deviceId byte) (*Readings, error) {
+func (mc *MeasurementCache) GetCurrent(deviceId byte) (*Readings, error) {
 	if item, ok := mc.items[deviceId]; ok {
 		if item.meter.GetState() == AVAILABLE {
 			return &item.readings.Current, nil
@@ -144,3 +145,52 @@ type ByteSlice []byte
 func (s ByteSlice) Len() int           { return len(s) }
 func (s ByteSlice) Less(i, j int) bool { return s[i] < s[j] }
 func (s ByteSlice) Swap(i, j int)      { s[i], s[j] = s[j], s[i] }
+
+type MeterReadings struct {
+	Historic ReadingSlice
+	Current  Readings
+	mux      sync.Mutex
+}
+
+func NewMeterReadings(devid uint8, maxAge time.Duration) *MeterReadings {
+	res := &MeterReadings{
+		Historic: ReadingSlice{},
+		Current: Readings{
+			UniqueId: fmt.Sprintf(UniqueIdFormat, devid),
+			DeviceId: devid,
+		},
+	}
+	go func(mr *MeterReadings) {
+		for {
+			time.Sleep(maxAge)
+			mr.mux.Lock()
+			mr.Historic = mr.Historic.NotOlderThan(time.Now().Add(-1 * maxAge))
+			mr.mux.Unlock()
+		}
+	}(res)
+	return res
+}
+
+func (mr *MeterReadings) Purge(devid uint8) {
+	mr.mux.Lock()
+	defer mr.mux.Unlock()
+
+	mr.Historic = ReadingSlice{}
+	mr.Current = Readings{
+		UniqueId: fmt.Sprintf(UniqueIdFormat, devid),
+		DeviceId: devid,
+	}
+}
+
+func (mr *MeterReadings) AddSnip(snip QuerySnip) {
+	mr.mux.Lock()
+	defer mr.mux.Unlock()
+
+	// 1. Merge the snip to the last values.
+	reading := mr.Current
+	reading.MergeSnip(snip)
+
+	// 2. store it
+	mr.Current = reading
+	mr.Historic = append(mr.Historic, reading)
+}
